# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Chat API
  version: 1.0.0

paths:
  /signup:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
                
        '400':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        default:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /mentions:
    get:
      summary: メンション未読通知を取得する
      description: 現在ログイン中のユーザーに対する、未読のメンション通知を返します
      tags:
        - Mentions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: メンション一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentionedMessages'
        '401':
          description: 認証エラー（トークン不備など）
        '500':
          description: サーバー内部エラー

  /users:
    get:
      summary: Get all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'    

  /messages:
    post:
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '201':
          description: Sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

    get:
      summary: Get messages with a user
      parameters:
        - name: room_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'

  /messages/{message_id}/read:
    post:
      summary: メッセージを既読として登録
      operationId: markMessageAsRead
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 登録成功
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /messages/unread:
    get:
      summary: Get unread messages for current user
      operationId: GetUnreadMessages
      responses:
        '200':
          description: 未読メッセージ一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'


  /messages/{message_id}/attachments:
    post:
      summary: メッセージに画像を添付する
      operationId: uploadMessageAttachment
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 添付ファイルのアップロード成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '400':
          description: リクエスト不正
        '404':
          description: メッセージが見つからない



  /chat-rooms:
    post:
      summary: Create or get one-on-one chat room
      operationId: ChatRoomsPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRoomInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - is_group
                  - created_at
                properties:
                  id:
                    type: integer
                  is_group:
                    type: boolean
                  created_at:
                    type: string
                    format: date-time
    get:
      summary: Get all chat rooms the user belongs to
      operationId: GetChatRooms
      responses:
        '200':
          description: List of chat rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatRoom'

  /chat-rooms/groups:
    post:
      summary: Create a new group chat room
      operationId: CreateGroupChatRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupChatInput'
      responses:
        '200':
          description: Group chat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatRoom'
  /me:
    get:
      operationId: GetMe
      summary: 現在ログインしているユーザーを取得
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'


components:
  schemas:
    UserInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          
    SignupResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
      required:
        - id
        - name
        - email

    User:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Human-readable error message

    MentionedMessage:
      type: object
      properties:
        message_id:
          type: integer
        content:
          type: string
        room_id:
          type: integer
        sender_id:
          type: integer
        created_at:
          type: string
          format: date-time

    MentionedMessages:
      type: array
      items:
        $ref: '#/components/schemas/MentionedMessage'

    MessageInput:
      type: object
      required:
        - room_id
        - text
      properties:
        room_id:
          type: integer
        text:
          type: string
        mention_usernames:
          type: array
          items:
            type: string
          nullable: true        # ← NULL許容（明示的に空でもOK）


    Message:
      type: object
      required:
        - id
        - sender_id
        - room_id
        - text
        - timestamp
      properties:
        id:
          type: integer
        sender_id:
          type: integer
        room_id:
          type: integer
        text:
          type: string
        timestamp:
          type: string
          format: date-time
        is_read:             
          type: boolean          
        attachments:         
          type: array
          items:
            $ref: '#/components/schemas/Attachment'

    Attachment:
      type: object
      properties:
        id:
          type: integer
        message_id:
          type: integer
        file_name:
          type: string
        url:
          type: string
        created_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    ChatRoomInput:
      type: object
      required:
        - target_user_id
      properties:
        target_user_id:
          type: integer

    CreateGroupChatInput:
      type: object
      required:
        - room_name
        - member_ids
      properties:
        room_name:
          type: string
        member_ids:
          type: array
          items:
            type: integer

    ChatRoom:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        room_name:
          type: string
          nullable: true
        is_group:
          type: boolean
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
